scripts/degrade_images/requirements.txt
git+https://github.com/NVlabs/ocrodeg@master#egg=ocrodeg
click
numpy
scipy
Pillow
